hubitat_agent/requirements-hubitat.txt
requests==2.31.0
psycopg2-binary==2.9.7
Flask==2.2.5
pytest==7.4.3